test:  删除react-native-switch测试demo引用components的方式
